Moved types and computed consts into definition fields
